# finsec-api/finsec_api/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FinSec Banking API",
    "description": "A secure banking API with authentication and financial management capabilities",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "support@example.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "paths": {
...
    "/api/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get transactions",
        "description": "Get transactions with filters",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "dateRange",
            "in": "query",
            "type": "string",
            "enum": ["all", "today", "week", "month"],
            "description": "Filter transactions by date range"
          },
          {
            "name": "type",
            "in": "query",
            "type": "string",
            "enum": ["all", "debit", "credit"],
            "description": "Filter transactions by type"
          },
          {
            "name": "minAmount",
            "in": "query",
            "type": "number",
            "description": "Minimum transaction amount"
          },
          {
            "name": "maxAmount",
            "in": "query",
            "type": "number",
            "description": "Maximum transaction amount"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string",
            "description": "Search in merchant name or category"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 10,
            "maximum": 50,
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "transactions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {"type": "string"},
                      "type": {"type": "string"},
                      "amount": {"type": "number"},
                      "date": {"type": "string"},
                      "category": {"type": "string"},
                      "merchant": {"type": "string"},
                      "status": {"type": "string"},
                      "paymentMethod": {"type": "string"}
                    }
                  }
                },
                "pagination": {
                  "type": "object",
                  "properties": {
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "limit": {"type": "integer"},
                    "pages": {"type": "integer"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/transactions/{transaction_id}": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get transaction details",
        "description": "Get detailed information for a specific transaction",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the transaction to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "id": {"type": "string"},
                "type": {"type": "string"},
                "amount": {"type": "number"},
                "date": {"type": "string"},
                "description": {"type": "string"},
                "category": {"type": "string"},
                "merchant": {"type": "string"},
                "reference": {"type": "string"},
                "status": {"type": "string"},
                "paymentMethod": {"type": "string"}
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    },
...
  }
} 